requirements.txt
streamlit
pandas
smtplib
python-dotenv
langchain_groq
